# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  requests.UpdateMerchantDocumentStatusRequest:
    properties:
      document_id:
        type: integer
      merchant_id:
        minimum: 1
        type: integer
      note:
        type: string
      status:
        type: string
    required:
    - merchant_id
    - note
    - status
    type: object
...
  response.ApiResponseMerchantDocument:
    properties:
      data:
        $ref: '#/definitions/response.MerchantDocumentResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantDocumentAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantDocumentDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.MerchantDocumentResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.MerchantDocumentResponse:
    properties:
      created_at:
        type: string
      document_type:
        type: string
      document_url:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      note:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  response.MerchantDocumentResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      document_type:
        type: string
      document_url:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      note:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant-document-command/create:
    post:
      consumes:
      - multipart/form-data
      description: Upload and create a new merchant document record
      parameters:
      - description: Merchant ID
        in: formData
        name: merchant_id
        required: true
        type: integer
      - description: Document type (e.g., NIB, SIUP)
        in: formData
        name: document_type
        required: true
        type: string
      - description: Document file
        in: formData
        name: document_url
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created merchant document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to create merchant document
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new merchant document
      tags:
      - Merchant Document Command
  /api/merchant-document-command/permanent/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a merchant document record by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted merchant document record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentDeleteAt'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to delete merchant document permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete merchant document
      tags:
      - Merchant Document Command
  /api/merchant-document-command/permanent/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all trashed merchant document records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all merchant documents permanently
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentAll'
        "500":
          description: Failed to delete merchant documents permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all trashed merchant documents
      tags:
      - Merchant Document Command
  /api/merchant-document-command/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed merchant document record by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored merchant document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to restore merchant document
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore trashed merchant document
      tags:
      - Merchant Document Command
  /api/merchant-document-command/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all trashed merchant document records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocumentAll'
        "500":
          description: Failed to restore merchant documents
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all trashed merchant documents
      tags:
      - Merchant Document Command
  /api/merchant-document-command/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Move a merchant document record to trash by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully moved merchant document to trash
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to move merchant document to trash
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Move merchant document to trash
      tags:
      - Merchant Document Command
  /api/merchant-document-command/update-status/{id}:
    post:
      consumes:
      - application/json
      description: Update the status and note of a merchant document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update status request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateMerchantDocumentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated document status
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to update document status
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Update merchant document status
      tags:
      - Merchant Document Command
  /api/merchant-document-command/update/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Update an existing merchant document record
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Merchant ID
        in: formData
        name: merchant_id
        required: true
        type: integer
      - description: Document type
        in: formData
        name: document_type
        required: true
        type: string
      - description: New document file
        in: formData
        name: document_url
        type: file
      - description: Old document URL if no new file is uploaded
        in: formData
        name: document_url_old
        type: string
      - description: Document status
        in: formData
        name: status
        required: true
        type: string
      - description: Document note
        in: formData
        name: note
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated merchant document
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to update merchant document
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Update merchant document
      tags:
      - Merchant Document Command
...
swagger: "2.0"